# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BhuSetu",
    "description": "BhuSetu: Intelligent Land Record Digitization and Validation System for SIH Problem Statement 26018 (Dept of Land Resources, Govt of India)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/ownership/chain/{document_id}": {
      "get": {
        "tags": [
          "Ownership"
        ],
        "summary": "Get Ownership Chain",
        "description": "Returns genealogical ownership graph, transfer timeline, and chain continuity analysis.",
        "operationId": "get_ownership_chain_api_ownership_chain__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/parcels/{ulpin}": {
      "get": {
        "tags": [
          "Parcels & ULPIN"
        ],
        "summary": "Get Parcel By Ulpin",
        "description": "Looks up a cadastral parcel by its 14-digit ULPIN (Bhu-Aadhaar).\nReturns GeoJSON polygon geometry, boundary attributes, and linked digitized record status.\nApplies PII masking on owner names for unauthorized roles.",
        "operationId": "get_parcel_by_ulpin_api_parcels__ulpin__get",
        "parameters": [
          {
            "name": "ulpin",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ulpin"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/parcels/": {
      "get": {
        "tags": [
          "Parcels & ULPIN"
        ],
        "summary": "List Parcels",
        "description": "Lists cadastral parcels in GeoJSON format, optionally filtered by village.",
        "operationId": "list_parcels_api_parcels__get",
        "parameters": [
          {
            "name": "village",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter parcels by village name",
              "title": "Village"
            },
            "description": "Filter parcels by village name"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ownership/chain/{document_id}": {
      "get": {
        "tags": [
          "Ownership"
        ],
        "summary": "Get Ownership Chain",
        "description": "Returns genealogical ownership graph, transfer timeline, and chain continuity analysis.",
        "operationId": "get_ownership_chain_api_v1_ownership_chain__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/parcels/{ulpin}": {
      "get": {
        "tags": [
          "Parcels & ULPIN"
        ],
        "summary": "Get Parcel By Ulpin",
        "description": "Looks up a cadastral parcel by its 14-digit ULPIN (Bhu-Aadhaar).\nReturns GeoJSON polygon geometry, boundary attributes, and linked digitized record status.\nApplies PII masking on owner names for unauthorized roles.",
        "operationId": "get_parcel_by_ulpin_api_v1_parcels__ulpin__get",
        "parameters": [
          {
            "name": "ulpin",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ulpin"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parcels/": {
      "get": {
        "tags": [
          "Parcels & ULPIN"
        ],
        "summary": "List Parcels",
        "description": "Lists cadastral parcels in GeoJSON format, optionally filtered by village.",
        "operationId": "list_parcels_api_v1_parcels__get",
        "parameters": [
          {
            "name": "village",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter parcels by village name",
              "title": "Village"
            },
            "description": "Filter parcels by village name"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}